Finished resize()
resize() creates a dummy pointer to point at the old array, then the pointer arr is updated to point to a new integer array in memory of size 2*maxCapacity. The new array is populated with the old arrays data and the old array at pointer tmp is freed from memory. maxCapacity is set to the new max.