viewItems() added, deleteIndex() and remove() implemented
viewItems() was created to assist with viewing the entire contents of the vector.

deleteIndex() is finished and will erase the item at the specified index while decrementing the number of items contained in the vector.

remove() will continually call find() until no more instances are found within the vector and call deleteIndex() on each index found.